initial distribution refactor
