+++ scvi/core/_log_likelihood.py
@@ -1,7 +1,7 @@
 """File for computing log likelihood of the data"""
 import numpy as np
 import torch
-import torch.nn.functional as F
+
 from torch import logsumexp
 from torch.distributions import Beta, Normal
 
@@ -216,135 +216,3 @@ def compute_marginal_log_likelihood_autozi(autozivae, data_loader, n_samples_mc=
     n_samples = len(data_loader.indices)
     # The minus sign is there because we actually look at the negative log likelihood
     return -log_lkl / n_samples
-
-
-def log_zinb_positive(x, mu, theta, pi, eps=1e-8):
-    """Note: All inputs are torch Tensors
-    log likelihood (scalar) of a minibatch according to a zinb model.
-    Notes:
-    We parametrize the bernoulli using the logits, hence the softplus functions appearing
-
-    Variables:
-    mu: mean of the negative binomial (has to be positive support) (shape: minibatch x genes)
-    theta: inverse dispersion parameter (has to be positive support) (shape: minibatch x genes)
-    pi: logit of the dropout parameter (real support) (shape: minibatch x genes)
-    eps: numerical stability constant
-
-    """
-
-    # theta is the dispersion rate. If .ndimension() == 1, it is shared for all cells (regardless of batch or labels)
-    if theta.ndimension() == 1:
-        theta = theta.view(
-            1, theta.size(0)
-        )  # In this case, we reshape theta for broadcasting
-
-    softplus_pi = F.softplus(-pi)  #  uses log(sigmoid(x)) = -softplus(-x)
-    log_theta_eps = torch.log(theta + eps)
-    log_theta_mu_eps = torch.log(theta + mu + eps)
-    pi_theta_log = -pi + theta * (log_theta_eps - log_theta_mu_eps)
-
-    case_zero = F.softplus(pi_theta_log) - softplus_pi
-    mul_case_zero = torch.mul((x < eps).type(torch.float32), case_zero)
-
-    case_non_zero = (
-        -softplus_pi
-        + pi_theta_log
-        + x * (torch.log(mu + eps) - log_theta_mu_eps)
-        + torch.lgamma(x + theta)
-        - torch.lgamma(theta)
-        - torch.lgamma(x + 1)
-    )
-    mul_case_non_zero = torch.mul((x > eps).type(torch.float32), case_non_zero)
-
-    res = mul_case_zero + mul_case_non_zero
-
-    return res
-
-
-def log_nb_positive(x, mu, theta, eps=1e-8):
-    """Note: All inputs should be torch Tensors
-    log likelihood (scalar) of a minibatch according to a nb model.
-
-    Variables:
-    mu: mean of the negative binomial (has to be positive support) (shape: minibatch x genes)
-    theta: inverse dispersion parameter (has to be positive support) (shape: minibatch x genes)
-    eps: numerical stability constant
-
-    """
-    if theta.ndimension() == 1:
-        theta = theta.view(
-            1, theta.size(0)
-        )  # In this case, we reshape theta for broadcasting
-
-    log_theta_mu_eps = torch.log(theta + mu + eps)
-
-    res = (
-        theta * (torch.log(theta + eps) - log_theta_mu_eps)
-        + x * (torch.log(mu + eps) - log_theta_mu_eps)
-        + torch.lgamma(x + theta)
-        - torch.lgamma(theta)
-        - torch.lgamma(x + 1)
-    )
-
-    return res
-
-
-def log_mixture_nb(x, mu_1, mu_2, theta_1, theta_2, pi, eps=1e-8):
-    """Note: All inputs should be torch Tensors
-    log likelihood (scalar) of a minibatch according to a mixture nb model.
-    pi is the probability to be in the first component.
-
-    For totalVI, the first component should be background.
-
-    Parameters
-    ----------
-    mu1: mean of the first negative binomial component (has to be positive support) (shape: minibatch x genes)
-    theta1: first inverse dispersion parameter (has to be positive support) (shape: minibatch x genes)
-    mu2: mean of the second negative binomial (has to be positive support) (shape: minibatch x genes)
-    theta2: second inverse dispersion parameter (has to be positive support) (shape: minibatch x genes)
-        If None, assume one shared inverse dispersion parameter.
-    eps: numerical stability constant
-
-
-    Returns
-    -------
-
-    """
-    if theta_2 is not None:
-        log_nb_1 = log_nb_positive(x, mu_1, theta_1)
-        log_nb_2 = log_nb_positive(x, mu_2, theta_2)
-    # this is intended to reduce repeated computations
-    else:
-        theta = theta_1
-        if theta.ndimension() == 1:
-            theta = theta.view(
-                1, theta.size(0)
-            )  # In this case, we reshape theta for broadcasting
-
-        log_theta_mu_1_eps = torch.log(theta + mu_1 + eps)
-        log_theta_mu_2_eps = torch.log(theta + mu_2 + eps)
-        lgamma_x_theta = torch.lgamma(x + theta)
-        lgamma_theta = torch.lgamma(theta)
-        lgamma_x_plus_1 = torch.lgamma(x + 1)
-
-        log_nb_1 = (
-            theta * (torch.log(theta + eps) - log_theta_mu_1_eps)
-            + x * (torch.log(mu_1 + eps) - log_theta_mu_1_eps)
-            + lgamma_x_theta
-            - lgamma_theta
-            - lgamma_x_plus_1
-        )
-        log_nb_2 = (
-            theta * (torch.log(theta + eps) - log_theta_mu_2_eps)
-            + x * (torch.log(mu_2 + eps) - log_theta_mu_2_eps)
-            + lgamma_x_theta
-            - lgamma_theta
-            - lgamma_x_plus_1
-        )
-
-    logsumexp = torch.logsumexp(torch.stack((log_nb_1, log_nb_2 - pi)), dim=0)
-    softplus_pi = F.softplus(-pi)
-
-    log_mixture_nb = logsumexp - softplus_pi
-
-    return log_mixture_nb
