+++ scvi/core/distributions/__init__.py
@@ -1,0 +1,11 @@
+from ._negative_binomial import (
+    NegativeBinomial,
+    NegativeBinomialMixture,
+    ZeroInflatedNegativeBinomial,
+)
+
+__all__ = [
+    "NegativeBinomial",
+    "NegativeBinomialMixture",
+    "ZeroInflatedNegativeBinomial",
+]
